sched: complete rework of context switching
The old strategy was to only do context switching
from within the PendSV handler.  This worked fine
until now because all syscalls were handled either
atomically or just returned -EAGAIN if the
resource was locked or busy.  However, with the
introduction of I/O wait, we need to be able to
sleep directly from within the kernel by moving
the context switching completely into the kernel.

+++ kernel/fs/read.c
@@ -9,9 +9,9 @@
 #include <stddef.h>
 #include <toolchain.h>
 
-ssize_t sys_read(int fd, __user void *buf, size_t len)
+long sys_read(int fd, __user void *buf, size_t len)
 {
-	ssize_t ret;
+	long ret;
 	void *copy;
 
 	struct file *f = file_get(fd);
